backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
django-cors-headers==4.4.0
psycopg[binary]==3.2.10
python-dotenv==1.0.1
gunicorn==23.0.0
whitenoise==6.7.0
